# perf: Use godirwalk in gitserver to speed up listing cloned repos